avoid warnings on VS2008
git-svn-id: https://svn.osgeo.org/gdal/trunk@22648 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/gml/gmlreader.cpp
@@ -1036,7 +1036,7 @@ int GMLReader::LoadClasses( const char *pszFile )
     }
 
     VSIFSeekL( fp, 0, SEEK_END );
-    nLength = VSIFTellL( fp );
+    nLength = (int) VSIFTellL( fp );
     VSIFSeekL( fp, 0, SEEK_SET );
 
     pszWholeText = (char *) VSIMalloc(nLength+1);
